+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+ * Work around in `Builder` the inability to pass empty arguments on Windows
+ * Catch more exceptions that can occur in `Loader` when caching resources ([pull #226](https://github.com/bytedeco/javacpp/pull/226))
  * Add `.a` as an allowed library extension for iOS so they can get bundled
  * Fix `Parser` failing on variables with direct list initialization `{ ... }` ([issue #223](https://github.com/bytedeco/javacpp/issues/223))
  * Allow `Parser` to map and cast function pointers to `Pointer`